% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_debug}
\alias{log_debug}
\title{Log a debug message}
\usage{
log_debug(message, ...)
}
\arguments{
\item{message}{your message to log}

\item{...}{more elements of the message, to be concatenated with the message}
}
\value{
invisibly returns TRUE/FALSE
}
\description{
Log messages will only be emitted if the log priority matches or is higher
than the priority of your message
}
\examples{
\dontrun{
log_debug("This is a debug message")
Sys.setenv("LOG_LEVEL" = "TRACE")
log_debug("This is a debug message")
}
}
