% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_error}
\alias{log_error}
\title{Log an error message}
\usage{
log_error(message)
}
\arguments{
\item{message}{your message to log}
}
\value{
invisibly returns TRUE/FALSE
}
\description{
Log messages will only be emitted if the log priority matches or is higher
than the priority of your message
}
\examples{
\dontrun{
log_error("This is an error message")
Sys.setenv("LOG_LEVEL" = "TRACE")
log_error("This is an error message")
}
}
