% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_fatal}
\alias{log_fatal}
\title{Log a fatal message}
\usage{
log_fatal(message)
}
\arguments{
\item{message}{your message to log}
}
\value{
invisibly returns TRUE/FALSE
}
\description{
Log messages will only be emitted if the log priority matches or is higher
than the priority of your message
}
\examples{
\dontrun{
log_fatal("This is a fatal message")
Sys.setenv("LOG_LEVEL" = "TRACE")
log_fatal("This is a fatal message")
}
}
