% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_info}
\alias{log_info}
\title{Log an info message}
\usage{
log_info(message, ...)
}
\arguments{
\item{message}{your message to log}

\item{...}{more elements of the message, to be concatenated with the message}
}
\value{
invisibly returns TRUE/FALSE
}
\description{
Log messages will only be emitted if the log priority matches or is higher
than the priority of your message
}
\examples{
\dontrun{
log_info("This is an info message")
Sys.setenv("LOG_LEVEL" = "TRACE")
log_info("This is an info message")
}
}
