% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_trace}
\alias{log_trace}
\title{Log a trace message}
\usage{
log_trace(message)
}
\arguments{
\item{message}{your message to log}
}
\value{
invisibly returns TRUE/FALSE
}
\description{
Log messages will only be emitted if the log priority matches or is higher
than the priority of your message
}
\examples{
\dontrun{
log_trace("This is a trace message")
Sys.setenv("LOG_LEVEL" = "TRACE")
log_trace("This is a trace message")
}
}
