% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_warn}
\alias{log_warn}
\title{Log a warning message}
\usage{
log_warn(message)
}
\arguments{
\item{message}{your message to log}
}
\value{
invisibly returns TRUE/FALSE
}
\description{
Log messages will only be emitted if the log priority matches or is higher
than the priority of your message
}
\examples{
\dontrun{
log_warn("This is a warning message")
Sys.setenv("LOG_LEVEL" = "TRACE")
log_warn("This is a warning message")
}
}
